Add optimized writer integration tests
